DROID-3375 fixes
