+++ localization/src/main/res/values/strings.xml
@@ -1981,5 +1981,6 @@ Please provide specific details of your needs here.</string>
     <string name="property_select_format_title">Select property format</string>
     <string name="property_edit_menu_delete">Delete from space</string>
     <string name="property_edit_menu_unlink">Unlink from type</string>
+    <string name="delete_space_checkbox_text">I have read and want to delete this space</string>
 
 </resources>
